kimi_k2_benchmark/pyproject.toml
[project]
name = "kimi-k2-benchmark"
version = "0.1.0"
description = "Comprehensive benchmark framework for Kimi K2 Heavy Mode vs Normal vs Qwen3-Coder"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Jose", email = "jose@rtx.local"}
]

dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "radon>=6.0.1",
    "tqdm>=4.65.0",
    "rich>=13.5.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "nbformat>=5.9.0",
    "ipykernel>=6.25.0",
    "pytest-mock>=3.11.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "tests*"]
